Tentative answer to Q7.
Had to Google/troubleshoot a lot. Still shakey with the use of increment but I'm getting there!
Also I think it'd be cool to update it so the inputs arent case sensitive. I looked into doing that and if it would be better for splice or for the change case function.